+++ src/DevBetterWeb.Web/Pages/Admin/Books/Index.cshtml
@@ -52,3 +52,8 @@
 }
     </tbody>
 </table>
+<div>
+    <a href="./Admin/Users">
+        Manage Users
+    </a>
+</div>
